requirements.txt
fastapi>=0.95.0
uvicorn>=0.18.0
psycopg2-binary>=2.9.6
python-dotenv>=0.21.0
requests>=2.26.0
passlib>=1.7.4
python-jose>=3.3.0
bcrypt>=3.2.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
